Renamed ReservationService, extracted repository methods. Cleaned up ReservationTransformer
