+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/transformers/ReservationTransformer.java
@@ -1,7 +1,6 @@
 package com.application.hotelbooking.transformers;
 
 import com.application.hotelbooking.domain.Reservation;
-import com.application.hotelbooking.domain.ReservationModel;
 import com.application.hotelbooking.dto.ReservationDTO;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,25 +15,17 @@ public class ReservationTransformer {
     @Autowired
     private ModelMapper modelMapper;
 
-    public ReservationModel transformToReservationModel(Reservation reservation){
-        return modelMapper.map(reservation, ReservationModel.class);
-    }
-
-    public Reservation transformToReservation(ReservationModel reservationModel){
-        return modelMapper.map(reservationModel, Reservation.class);
-    }
-
-    public List<ReservationModel> transformToReservationModels(List<Reservation> reservations){
-        return reservations.stream()
-                .map(reservation -> modelMapper.map(reservation, ReservationModel.class))
-                .collect(Collectors.toList());
-    }
-
     public ReservationDTO transformToReservationDTO(Reservation reservation){
         return modelMapper.map(reservation, ReservationDTO.class);
     }
 
     public Reservation transformToReservation(ReservationDTO reservationDTO){
         return modelMapper.map(reservationDTO, Reservation.class);
+    }
+
+    public List<ReservationDTO> transformToReservationDTOs(List<Reservation> reservations){
+        return reservations.stream()
+                .map(reservation -> modelMapper.map(reservation, ReservationDTO.class))
+                .collect(Collectors.toList());
     }
 }
